requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
line-bot-sdk==3.8.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
yfinance==0.2.36
python-dotenv==1.0.0
pydantic==2.5.0
alembic==1.13.1
requests==2.31.0
python-dateutil==2.8.2
